refactor: improve integration testing
