+++ tests/integration/local/test_api_key.py
@@ -1,43 +1,48 @@
 import hubble
 import requests
-from tests.integration.bff.conftest import (
+
+from now.constants import ACCESS_PATHS
+from tests.integration.local.conftest import (
     BASE_URL,
-    HOST,
-    PORT,
     SEARCH_URL,
     get_flow,
-    index_data,
+    get_request_body,
 )
-from tests.integration.test_end_to_end import assert_search
-
-from now.admin.utils import get_default_request_body
+from tests.integration.remote.test_end_to_end import assert_search
 
 API_KEY = 'my_key'
 update_api_keys_url = f'{BASE_URL}/admin/updateApiKeys'
 update_emails_url = f'{BASE_URL}/admin/updateUserEmails'
 
 
-def get_request_body():
-    request_body = get_default_request_body(host=HOST, secured=True)
-    request_body['port'] = PORT
-    return request_body
-
-
-def test_add_key(start_bff, setup_service_running, random_index_name, tmpdir):
+def test_add_key(
+    simple_data, start_bff, setup_service_running, random_index_name, tmpdir
+):
     client = hubble.Client(
-        token=get_request_body()['jwt']['token'], max_retries=None, jsonify=True
+        token=get_request_body(secured=True)['jwt']['token'],
+        max_retries=None,
+        jsonify=True,
     )
     admin_email = client.get_user_info()['data'].get('email')
 
     f = get_flow(
         preprocessor_args={'admin_emails': [admin_email]},
-        indexer_args={'admin_emails': [admin_email], 'index_name': random_index_name},
+        indexer_args={
+            'admin_emails': [admin_email],
+            'index_name': random_index_name,
+            'document_mappings': [['encoderclip', 512, ['title']]],
+        },
         tmpdir=tmpdir,
     )
     with f:
-        index_data(f, jwt=get_request_body()['jwt'])
-
-        request_body = get_request_body()
+        f.index(
+            simple_data,
+            parameters={
+                'access_paths': ACCESS_PATHS,
+                'jwt': get_request_body(secured=True)['jwt'],
+            },
+        )
+        request_body = get_request_body(secured=True)
         # Test adding user email
         request_body['user_emails'] = ['florian.hoenicke@jina.ai']
         response = requests.post(
@@ -48,7 +53,7 @@ def test_add_key(start_bff, setup_service_running, random_index_name, tmpdir):
 
         # test api keys
         # search with invalid api key
-        request_body = get_request_body()
+        request_body = get_request_body(secured=True)
         request_body['query'] = [
             {'name': 'text', 'value': 'girl on motorbike', 'modality': 'text'}
         ]
@@ -57,7 +62,7 @@ def test_add_key(start_bff, setup_service_running, random_index_name, tmpdir):
         request_body['limit'] = 9
         assert_search(SEARCH_URL, request_body, expected_status_code=500)
         print('# add api key')
-        request_body_update_keys = get_request_body()
+        request_body_update_keys = get_request_body(secured=True)
         request_body_update_keys['api_keys'] = [API_KEY]
         response = requests.post(
             update_api_keys_url,
